+++ source/Information/FileStructures_for_HHD_Hex_Editor_Neo/philips_mgr_chan_s_fta.db.h
@@ -13,43 +13,30 @@ struct SHeader
   uint32 channelBlockSize;
 };
 
-struct SChannel
+struct SChannel_fta
 {
 	uint32 curProgNr;
 	uint32 u1;
 	uint8 u2[8];
 	uint32 favNr;
-	union
-	{
-		char chName1[200];
-#pragma byte_order (BigEndian)
-		big_endian wchar_t chName2[100];
-#pragma byte_order ()
-		struct 
-		{
-			uint8 zero;
-			wchar_t chName3[99];
-			uint8 zero2;
-		} chName4;
-	} chName;
+    char chName1[200];
 	uint16 u3;
 	uint8 u3b[208];
 	uint8 u3c[2];
 	uint16 u3d;	
 	uint8 u4[10];
-	uint16 freqInMhz1;
-	uint16 u5;
+	uint32 freqInMhz1;
+	//uint16 u5;
 	uint16 u6;
 	uint16 symRate;
-	uint32 curProgNr2;
-	uint32 prevProgNr;
+	uint32 oldProgNr;
+	uint32 channelIndex;
 	uint16 tsid;
 	uint16 u7;
 	uint16 sid;
 	uint16 onid;
 	uint16 freqInMhz2;
-	uint16 u9;
-	uint32 u10;
+	uint8 padding[6];
 };
 
 struct SFooter
@@ -70,8 +57,105 @@ public struct Philips_mgr_chan_s_fta
 	
 	SHeader header;
 
-	var recordCount = header.channelBlockSize / sizeof(SChannel);
-    SChannel channels[recordCount];
+	var recordCount = header.channelBlockSize / sizeof(SChannel_fta);
+    SChannel_fta channels[recordCount];
+	
+	SFooter footer;	
+};
+
+//#########################################################
+
+struct SChannel_pkg
+{
+	uint32 curProgNr;
+	uint32 u1;
+	uint8 u2[8];
+	uint32 favNr;
+    char chName1[200];
+	uint16 u3;
+	uint8 u3b[208];
+	uint8 u3c[2];
+	uint16 u3d;	
+	uint8 u4[10];
+	uint32 freqInMhz1;
+	//uint16 u5;
+	uint16 u6;
+	uint16 symRate;
+	uint32 oldProgNr;
+	uint32 channelIndex;
+	uint16 tsid;
+	uint16 u7;
+	uint16 sid;
+	uint16 onid;
+	uint16 freqInMhz2;
+	uint8 padding[10];
+};
+
+public struct Philips_mgr_chan_s_pkg
+{
+    var docSize = GetDocumentSize();
+
+    char filename[32];
+	
+	SHeader header;
+
+	var recordCount = header.channelBlockSize / sizeof(SChannel_pkg);
+    SChannel_pkg channels[recordCount];
+	
+	SFooter footer;	
+};
+
+//#########################################################
+
+struct CChannel
+{
+	uint32 curProgNr;
+	uint32 u1;
+	uint8 u2[8];
+	uint32 favNr;
+	union
+	{
+		char chName1[200];
+#pragma byte_order (BigEndian)
+		big_endian wchar_t chName2[100];
+#pragma byte_order ()
+		struct 
+		{
+			uint8 zero;
+			wchar_t chName3[99];
+			uint8 zero2;
+		} chName4;
+	} chName;
+	uint16 u3;
+	uint16 u3b;
+	char provider[200];
+	uint8 u4[16];
+	uint32 freqInHz;
+	uint16 u6;
+	uint16 not_symRate;
+	uint32 oldProgNr;
+	uint8 u7[4];
+	uint32 channelIndex;
+	uint16 tsid;
+	uint16 symRate_maybe;
+	uint16 sid;
+	uint16 onid;
+	//uint16 freqInMhz2;
+	//uint16 u9;
+	uint32 u10;
+};
+
+
+public struct Philips_mgr_chan_dvbt
+{
+    var docSize = GetDocumentSize();
+
+    char filename[32];
+	
+	SHeader header;
+
+	var recordCount = header.channelBlockSize / sizeof(CChannel);
+    CChannel channels[recordCount];
 	
 	SFooter footer;	
 };